+++ netmiko/extreme/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import unicode_literals
-from netmiko.extreme.extreme import ExtremeSSH
-from netmiko.extreme.extreme import ExtremeTelnet
+from netmiko.extreme.extreme_exos import ExtremeSSH
+from netmiko.extreme.extreme_exos import ExtremeTelnet
 from netmiko.extreme.extreme_wing_ssh import ExtremeWingSSH
 
 __all__ = ['ExtremeSSH', 'ExtremeWingSSH', 'ExtremeTelnet']
